+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v2.1.0 - 2025-02-XX
+
+### Added
+
+- [CAP Queue] Add support for defining successor and failed events of event handlers. See documentation for how to use it.
+- [CAP Queue] Allow to propagate cds.context properties (e.g. features). This can be configured per event (`cds.env.requires[<SERVICE>].queued.propagatedContextProperties = ["features"]`)
+
+### Fixed
+
+- The event property `retryOpenAfter` now works as expected.
+
 ## v2.0.4 - 2025-01-22
 
 ### Fixed